+++ changelog.md
@@ -1,3 +1,17 @@
+2009.4.51
+---------
+
+### Feature
+
+* minimal dependency ( used to depends on hack-contrib / mps, not now )
+
+2009.4.50
+---------
+
+### Fix
+
+* use show_status_message in Utils
+
 2009.4.30
 -----------